Move getPowersForIngredients to mechanics
